titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/bank-deposit/data-access-bank-deposit/src/lib/+state/bank-deposit.selectors.ts
import { createSelector, createFeatureSelector } from '@ngrx/store';
import {
  BankDepositState,
  BANK_DEPOSIT_FEATURE_KEY
} from './bank-deposit.state';

export const selectBankDepositState = createFeatureSelector<BankDepositState>(
  BANK_DEPOSIT_FEATURE_KEY
);

const selectIsLoading = createSelector(
  selectBankDepositState,
  state => state.isLoading
);

const selectBankDepositList = createSelector(
  selectBankDepositState,
  state => state.bankDepositData
);

const selectError = createSelector(
  selectBankDepositState,
  state => state.error
);

const selectTransactionDetails = createSelector(
  selectBankDepositState,
  state => state.transacionDetails
);

export const BankDepositSelectors = {
  selectError,
  selectIsLoading,
  selectBankDepositList,
  selectTransactionDetails
};
